Refactored the naming.
